+++ src/api/plugins/tdbp/tests/conftest.py
@@ -6,6 +6,7 @@
 import pytest
 from warren.tests.fixtures.app import http_client
 from warren.tests.fixtures.asynchronous import anyio_backend
+from warren.tests.fixtures.auth import auth_headers
 from warren.tests.fixtures.db import (
     db_engine,
     db_session,