+++ src/api/plugins/tdbp/tests/test_api.py
@@ -1,46 +1,44 @@
 """Tests for the TdBP Warren plugin."""
-from datetime import date, datetime
+
+from datetime import datetime
 
 import httpx
 import pytest
 from pydantic import ValidationError
+from warren.utils import LTIUser, forge_lti_token
 
 from warren_tdbp.models import Grades, Scores, SlidingWindow
 
 
 @pytest.mark.anyio
-@pytest.mark.parametrize(
-    "query_params",
-    [
-        {},  # Missing parameters
-        {"course_id": 1234},  # Wrong "course_id" parameter data type
-        {
-            "course_id": "course_id_101",
-            "until": "today",
-        },  # Wrong "until" parameter data type
-        {"until": date.today()},  # Missing "course_id" parameter
-        {"foo": "fake"},  # Unsupported parameter
-    ],
-)
 async def test_api_sliding_window_with_invalid_params(
-    query_params, http_client: httpx.AsyncClient
+    http_client: httpx.AsyncClient, auth_headers: dict
 ):
-    """Test '/window' indicator with invalid query parameters returns 422 HTTP code."""
-    response = await http_client.get("/api/v1/tdbp/window", params=query_params)
+    """Test `/window` endpoint with invalid query parameters returns 422 HTTP code."""
+    query_params = {"until": "today"}
+    response = await http_client.get(
+        "/api/v1/tdbp/window",
+        params=query_params,
+        headers=auth_headers,
+    )
 
     assert response.status_code == 422
 
 
 @pytest.mark.anyio
 async def test_api_sliding_window_with_valid_params(
-    http_client: httpx.AsyncClient, db_session, sliding_window_fake_dataset
+    http_client: httpx.AsyncClient,
+    db_session,
+    sliding_window_fake_dataset,
 ):
-    """Test '/window' indicator with query parameters returns correct output."""
-    course_id = "https://fake-lms.com/course/tdbp_101"
+    """Test `/window` endpoint with query parameters returns correct output."""
+    token = forge_lti_token(course_id="https://fake-lms.com/course/tdbp_101")
     date_until = datetime.now().date()
 
     response = await http_client.get(
-        "/api/v1/tdbp/window", params={"course_id": course_id, "until": date_until}
+        "/api/v1/tdbp/window",
+        params={"until": date_until},
+        headers={"Authorization": f"Bearer {token}"},
     )
 
     assert response.status_code == 200
@@ -52,38 +50,50 @@ async def test_api_sliding_window_with_valid_params(
 
 
 @pytest.mark.anyio
-@pytest.mark.parametrize(
-    "query_params",
-    [
-        {},  # Missing parameters
-        {"course_id": 1234},  # Wrong "course_id" parameter data type
-        {
-            "course_id": "course_id_101",
-            "until": "today",
-        },  # Wrong "until" parameter data type
-        {"until": date.today()},  # Missing "course_id" parameter
-        {"foo": "fake"},  # Unsupported parameter
-    ],
-)
-async def test_api_cohort_with_invalid_params(
-    query_params, http_client: httpx.AsyncClient
+async def test_api_sliding_window_with_invalid_auth_headers(
+    http_client: httpx.AsyncClient,
 ):
-    """Test '/cohort' indicator with invalid query parameters returns 422 HTTP code."""
-    response = await http_client.get("/api/v1/tdbp/cohort", params=query_params)
+    """Test `/window` endpoint with an invalid `auth_headers`."""
+    date_until = datetime.now().date()
+    response = await http_client.get(
+        "/api/v1/tdbp/window",
+        params={"until": date_until},
+        headers={"Authorization": "Bearer Wrong_Token"},
+    )
+
+    assert response.status_code == 401
+    assert response.json().get("detail") == "Could not validate credentials"
+
+
+@pytest.mark.anyio
+async def test_api_cohort_with_invalid_params(
+    http_client: httpx.AsyncClient, auth_headers: dict
+):
+    """Test `/cohort` endpoint with invalid query parameters returns 422 HTTP code."""
+    query_params = {"until": "today"}
+    response = await http_client.get(
+        "/api/v1/tdbp/cohort",
+        params=query_params,
+        headers=auth_headers,
+    )
 
     assert response.status_code == 422
 
 
 @pytest.mark.anyio
 async def test_api_cohort_with_valid_params(
-    http_client: httpx.AsyncClient, db_session, sliding_window_fake_dataset
+    http_client: httpx.AsyncClient,
+    db_session,
+    sliding_window_fake_dataset,
 ):
-    """Test '/window' indicator with query parameters returns correct output."""
-    course_id = "https://fake-lms.com/course/tdbp_101"
+    """Test `/window` endpoint with query parameters returns correct output."""
+    token = forge_lti_token(course_id="https://fake-lms.com/course/tdbp_101")
     date_until = datetime.now().date()
 
     response = await http_client.get(
-        "/api/v1/tdbp/cohort", params={"course_id": course_id, "until": date_until}
+        "/api/v1/tdbp/cohort",
+        params={"until": date_until},
+        headers={"Authorization": f"Bearer {token}"},
     )
 
     assert response.status_code == 200
@@ -92,106 +102,231 @@ async def test_api_cohort_with_valid_params(
 
 
 @pytest.mark.anyio
+async def test_api_cohort_with_invalid_auth_headers(
+    http_client: httpx.AsyncClient,
+):
+    """Test `/cohort` endpoint with an invalid `auth_headers`."""
+    date_until = datetime.now().date()
+    response = await http_client.get(
+        "/api/v1/tdbp/cohort",
+        params={"until": date_until},
+        headers={"Authorization": "Bearer Wrong_Token"},
+    )
+
+    assert response.status_code == 401
+    assert response.json().get("detail") == "Could not validate credentials"
+
+
+@pytest.mark.anyio
 @pytest.mark.parametrize(
     "query_params",
     [
-        {},  # Missing parameters
-        {"course_id": 1234},  # Wrong "course_id" parameter data type
         {
-            "course_id": "course_id_101",
             "until": "today",
         },  # Wrong "until" parameter data type
         {
-            "course_id": "course_id_101",
-            "student_id": 12345,
-        },  # Wrong "student_id" parameter data type
-        {
-            "course_id": "course_id_101",
-            "totals": 1,
+            "totals": 2,
         },  # Wrong "totals" parameter data type
         {
-            "course_id": "course_id_101",
-            "average": 1,
+            "average": 2,
         },  # Wrong "average" parameter data type
-        {"until": date.today()},  # Missing required "course_id" parameter
-        {"foo": "fake"},  # Unsupported parameter
     ],
 )
 async def test_api_scores_with_invalid_params(
-    query_params, http_client: httpx.AsyncClient
+    query_params, http_client: httpx.AsyncClient, auth_headers
 ):
-    """Test '/scores' indicator with invalid query parameters returns 422 HTTP code."""
-    response = await http_client.get("/api/v1/tdbp/scores", params=query_params)
+    """Test `/scores` endpoint with invalid query parameters returns 422 HTTP code."""
+    response = await http_client.get(
+        "/api/v1/tdbp/scores",
+        params=query_params,
+        headers=auth_headers,
+    )
 
     assert response.status_code == 422
 
 
 @pytest.mark.anyio
-async def test_api_scores_with_valid_params(
-    http_client: httpx.AsyncClient, db_session, sliding_window_fake_dataset
+async def test_api_scores_with_valid_params_instructor(
+    http_client: httpx.AsyncClient,
+    db_session,
+    sliding_window_fake_dataset,
 ):
-    """Test '/scores' indicator with query parameters returns correct output."""
-    course_id = "https://fake-lms.com/course/tdbp_101"
+    """Test `/scores` endpoint for an instructor returns correct output."""
+    token = forge_lti_token(
+        roles=("instructor",),
+        course_id="https://fake-lms.com/course/tdbp_101",
+    )
     date_until = datetime.now().date()
 
     response = await http_client.get(
-        "/api/v1/tdbp/scores", params={"course_id": course_id, "until": date_until}
+        "/api/v1/tdbp/scores",
+        params={"until": date_until, "average": True, "totals": True},
+        headers={"Authorization": f"Bearer {token}"},
     )
 
     assert response.status_code == 200
 
     try:
-        Scores.parse_obj(response.json())
+        scores = Scores.parse_obj(response.json())
     except ValidationError as err:
         pytest.fail(f"Scores indicator is invalid: {err}")
+
+    # Instructors can see average and totals scores
+    assert len(scores.scores) > 1
+    assert scores.average
+    assert scores.total
+
+
+@pytest.mark.anyio
+async def test_api_scores_with_valid_params_student(
+    http_client: httpx.AsyncClient,
+    db_session,
+    sliding_window_fake_dataset,
+):
+    """Test `/scores` endpoint for a student returns correct output."""
+    token = forge_lti_token(
+        user=LTIUser(
+            id="johndoe",
+            email="johndoe@example.com",
+        ),
+        roles=("student",),
+        course_id="https://fake-lms.com/course/tdbp_101",
+    )
+    date_until = datetime.now().date()
+
+    response = await http_client.get(
+        "/api/v1/tdbp/scores",
+        params={"until": date_until, "average": 1, "total": 1},
+        headers={"Authorization": f"Bearer {token}"},
+    )
+
+    assert response.status_code == 200
+
+    try:
+        scores = Scores.parse_obj(response.json())
+    except ValidationError as err:
+        pytest.fail(f"Scores indicator is invalid: {err}")
+
+    # Student can only see its scores, and cannot see total and average scores
+    scores.scores.pop("johndoe")
+    assert not scores.scores
+    assert not scores.average
+    assert not scores.total
+
+
+@pytest.mark.anyio
+async def test_api_scores_with_invalid_auth_headers(
+    http_client: httpx.AsyncClient,
+):
+    """Test `/scores` endpoint with an invalid `auth_headers`."""
+    date_until = datetime.now().date()
+    response = await http_client.get(
+        "/api/v1/tdbp/scores",
+        params={"until": date_until},
+        headers={"Authorization": "Bearer Wrong_Token"},
+    )
+
+    assert response.status_code == 401
+    assert response.json().get("detail") == "Could not validate credentials"
 
 
 @pytest.mark.anyio
 @pytest.mark.parametrize(
     "query_params",
     [
-        {},  # Missing parameters
-        {"course_id": 1234},  # Wrong "course_id" parameter data type
         {
-            "course_id": "course_id_101",
             "until": "today",
         },  # Wrong "until" parameter data type
         {
-            "course_id": "course_id_101",
-            "student_id": 12345,
-        },  # Wrong "student_id" parameter data type
-        {
-            "course_id": "course_id_101",
-            "average": 1,
+            "average": 2,
         },  # Wrong "average" parameter data type
-        {"until": date.today()},  # Missing required "course_id" parameter
-        {"foo": "fake"},  # Unsupported parameter
     ],
 )
 async def test_api_grades_with_invalid_params(
-    query_params, http_client: httpx.AsyncClient
+    query_params, http_client: httpx.AsyncClient, auth_headers: dict
 ):
-    """Test '/grades' indicator with invalid query parameters returns 422 HTTP code."""
-    response = await http_client.get("/api/v1/tdbp/grades", params=query_params)
+    """Test `/grades` endpoint with invalid query parameters returns 422 HTTP code."""
+    response = await http_client.get(
+        "/api/v1/tdbp/grades",
+        params=query_params,
+        headers=auth_headers,
+    )
 
     assert response.status_code == 422
 
 
 @pytest.mark.anyio
-async def test_api_grades_with_valid_params(
-    http_client: httpx.AsyncClient, db_session, sliding_window_fake_dataset
+async def test_api_grades_with_valid_params_instructor(
+    http_client: httpx.AsyncClient,
+    db_session,
+    sliding_window_fake_dataset,
 ):
-    """Test '/grades' indicator with query parameters returns correct output."""
-    course_id = "https://fake-lms.com/course/tdbp_101"
+    """Test `/grades` endpoint for an instructor returns correct output."""
+    token = forge_lti_token(
+        roles=("instructor",),
+        course_id="https://fake-lms.com/course/tdbp_101",
+    )
     date_until = datetime.now().date()
 
     response = await http_client.get(
-        "/api/v1/tdbp/grades", params={"course_id": course_id, "until": date_until}
+        "/api/v1/tdbp/grades",
+        params={"until": date_until, "average": True},
+        headers={"Authorization": f"Bearer {token}"},
     )
 
     assert response.status_code == 200
 
     try:
-        Grades.parse_obj(response.json())
+        grades = Grades.parse_obj(response.json())
     except ValidationError as err:
         pytest.fail(f"Grades indicator is invalid: {err}")
+
+    assert len(grades.grades) > 1
+    assert grades.average
+
+
+@pytest.mark.anyio
+async def test_api_grades_with_valid_params_student(
+    http_client: httpx.AsyncClient,
+    db_session,
+    sliding_window_fake_dataset,
+):
+    """Test `/grades` endpoint for a student returns correct output."""
+    token = forge_lti_token(
+        user=LTIUser(id="student_1", email="student_1@example.com"),
+        roles=("student",),
+        course_id="https://fake-lms.com/course/tdbp_101",
+    )
+    date_until = datetime.now().date()
+
+    response = await http_client.get(
+        "/api/v1/tdbp/grades",
+        params={"until": date_until},
+        headers={"Authorization": f"Bearer {token}"},
+    )
+
+    assert response.status_code == 200
+
+    try:
+        grades = Grades.parse_obj(response.json())
+    except ValidationError as err:
+        pytest.fail(f"Grades indicator is invalid: {err}")
+
+    # Student can only see its grades
+    assert len(grades.grades) == 1
+
+
+@pytest.mark.anyio
+async def test_api_grades_with_invalid_auth_headers(
+    http_client: httpx.AsyncClient,
+):
+    """Test `/grades` endpoint with an invalid `auth_headers`."""
+    date_until = datetime.now().date()
+    response = await http_client.get(
+        "/api/v1/tdbp/grades",
+        params={"until": date_until},
+        headers={"Authorization": "Bearer Wrong_Token"},
+    )
+
+    assert response.status_code == 401
+    assert response.json().get("detail") == "Could not validate credentials"