+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.3.0] - 2024-06-27
 
 ### Changed
 
@@ -28,6 +28,7 @@ and this project adheres to
 - Add a custom API Docker image, based on `warren:api-core-0.1.0`, for the tdbp plugin indicator
 - Add a custom APP Docker image, based on `warren:app-0.1.0`, to serve the Warren-TdBP frontend application
 
-[unreleased]: https://github.com/apui-avignon-university/warren-tdbp/compare/v0.2.0...main
+[unreleased]: https://github.com/apui-avignon-university/warren-tdbp/compare/v0.3.0...main
+[0.3.0]: https://github.com/apui-avignon-university/warren-tdbp/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/apui-avignon-university/warren-tdbp/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/apui-avignon-university/warren-tdbp/compare/dd18c21...v0.1.0